API key in block rather than settings page
