+++ src/js/EditorBlock.js
@@ -4,16 +4,14 @@
 import '../css/style.scss'
 import blockIcons from './icons.js'
 import getMapHTML from './getMapHTML.js'
-import formFields from './formFields.js'
 
 /**
  * Get WordPress libraries from the wp global
  */
 const { __ } = wp.i18n;
-// const { TextControl  } = wp.blocks.InspectorControls;
-// const { Component } = wp.element;
 const { Button } = wp.components;
-const { TextControl } = wp.blocks.InspectorControls;
+const { InspectorControls } = wp.blocks;
+const { TextControl, ToggleControl, RangeControl, SelectControl } = InspectorControls;
 const { Component } = wp.element;
 
 export default class EditorBlock extends Component {
@@ -21,36 +19,30 @@ export default class EditorBlock extends Component {
         super( ...arguments );
 
         this.saveApiKey = this.saveApiKey.bind( this );
-        this.clearApiKey = this.clearApiKey.bind( this );
 
         this.state = {
             apiKey: '',
             isSavedKey: false,
-            isLoading: true
+            isLoading: true,
+            isSaving: false,
+            keySaved: false,
         };
 
         const model = new wp.api.models.Settings();
         model.fetch().then( response => {
             this.setState({ apiKey: response.pantheon_google_map_block_api_key });
             if ( this.state.apiKey && this.state.apiKey !== '' ) {
-                this.setState({ isSavedKey: true, isLoading: false });
+                this.setState({ isSavedKey: true });
             }
+            this.setState({ isLoading: false });
         });
     }
 
     saveApiKey() {
-        if( this.state.apiKey !== '' ) {
-            const model = new wp.api.models.Settings({ pantheon_google_map_block_api_key:apiKey });
-            model.save().then( response => {
-                this.setState({ isSavedKey: true, isLoading: false });
-            });
-        }
-    }
-
-    clearApiKey() {
-        const model = new wp.api.models.Settings({ pantheon_google_map_block_api_key:'' });
+        this.setState({ isSaving: true });
+        const model = new wp.api.models.Settings({ pantheon_google_map_block_api_key: this.state.apiKey });
         model.save().then( response => {
-            this.setState({ isSavedKey: false, apiKey: '' });
+            this.setState({ isSavedKey: true, isLoading: false, isSaving: false, keySaved: true });
         });
     }
 
@@ -58,7 +50,20 @@ export default class EditorBlock extends Component {
         const { attributes, className, focus, setAttributes, setFocus } = this.props;
         const { location, mapType, zoom, interactive, maxWidth, maxHeight, aspectRatio } = attributes;
         const editorPadding = '0 1em';
-        const classNames = ( ! interactive ) ? `${className} ratio${aspectRatio}` : `${className}  ratio${aspectRatio} interactive`
+        const classNames = ( ! interactive ) ? `${className} ratio${aspectRatio}` : `${className}  ratio${aspectRatio} interactive`;
+
+        const linkOptions = [
+            {value: 'roadmap', label: __( 'roadmap' ) },
+            {value: 'satellite', label: __( 'satellite' ) },
+        ];
+
+        const aspectRatioOptions = [
+            {value: '2_1', label: __( '2:1' ) },
+            {value: '1_1', label: __( '1:1' ) },
+            {value: '4_3', label: __( '4:3' ) },
+            {value: '16_9', label: __( '16:9' ) },
+            {value: '1_2', label: __( '1:2' ) },
+        ];
 
         if ( !! this.state.isLoading  ) {
             return (
@@ -70,38 +75,94 @@ export default class EditorBlock extends Component {
             )
         }
 
+        const keyInput = (
+            <div>
+                <p style={{textAlign: 'center'}}>
+                    {__( 'A Google Maps API key is required, please enter one below.')  }
+                </p>
+                <TextControl
+                    key="api-input"
+                    value={ this.state.apiKey }
+                    onChange={ value => this.setState({ apiKey: value }) }
+                />
+                <p style={{textAlign: 'center', paddingBottom: '1em'}}>
+                    {__('Need an API key? Get one')}&nbsp;
+                    <a href="https://console.developers.google.com/flows/enableapi?apiid=maps_backend,static_maps_backend,maps_embed_backend&keyType=CLIENT_SIDE&reusekey=true">
+                        {__('here.')}
+                    </a><br />
+                    <Button 
+                        isPrimary 
+                        onClick={ this.saveApiKey }
+                        isBusy={ this.state.isSaving }
+                        disabled={ this.state.apiKey === '' }
+                    >
+                        {__('Save API key')}
+                    </Button>
+                </p>
+            </div>
+        );
+
         if ( ! this.state.isSavedKey  ) {
             return (
                 <div className={`${className} error`} style={{padding: editorPadding}}>
-                    <p style={{textAlign: 'center'}}>
-                        {__( 'A Google Maps API key is required, please enter one below.')  }
-                    </p>
-                    <TextControl
-                        key="api-input"
-                        value={ this.state.apiKey }
-                        onChange={ value => this.setState({ apiKey: value }) }
-                    />
-                    <p style={{textAlign: 'center', paddingBottom: '1em'}}>
-                        {__('Need an API key? Get one')}&nbsp;
-                        <a href="https://console.developers.google.com/flows/enableapi?apiid=maps_backend,static_maps_backend,maps_embed_backend&keyType=CLIENT_SIDE&reusekey=true">
-                            {__('here.')}
-                        </a><br />
-                        <Button onClick={ this.saveApiKey }>
-                            {__('Save API key')}
-                        </Button>
-                    </p>
+                    {keyInput}
                 </div>
-            );
+            )
         }
 
         return [
 			!! focus && (
-                <formFields 
-                    saveApiKey={this.saveApiKey}
-                    setState={this.setState}
-                    apiKey={this.state.apiKey}
-                    {...this.props}
-                />
+                <InspectorControls>
+                    {keyInput}
+                    {!! interactive ? (
+                        <SelectControl
+                            label={ __( 'Aspect Ratio' ) } 
+                            select={aspectRatio} 
+                            options={aspectRatioOptions} 
+                            onChange={ ( value ) => setAttributes( { aspectRatio: value } ) } 
+                            value={ aspectRatio }
+                        />
+                    ) : null}
+                    <RangeControl
+                        label={ __( 'Zoom Level' ) }
+                        value={ zoom }
+                        onChange={ ( value ) => setAttributes( { zoom: value } ) }
+                        min={ 5 }
+                        max={ 20 }
+                    />
+                    <SelectControl
+                        label={ __( 'Map Type' ) } 
+                        select={mapType} 
+                        options={linkOptions} 
+                        onChange={ ( value ) => setAttributes( { mapType: value } ) } 
+                        value={ mapType }
+                    />
+                    <ToggleControl
+                        label={ __( 'Toggle interactive map (on) or static image (off)' ) }
+                        checked={ !! interactive }
+                        onChange={ () => setAttributes( { interactive: ! interactive } ) }
+                    />
+                    {! interactive ? (
+                        <div>
+                            <TextControl 
+                                label={ __( 'Maximum width (in pixels)' ) } 
+                                onChange={ ( value ) => setAttributes( { maxWidth: Number.parseInt( value, 10 ) } ) }
+                                value={maxWidth}
+                                type='number'
+                                min={0}
+                                step={1}
+                            />
+                            <TextControl 
+                                label={ __( 'Maximum height (in pixels)' ) } 
+                                onChange={ ( value ) => setAttributes( { maxHeight: Number.parseInt( value, 10 ) } ) }
+                                value={maxHeight}
+                                type='number'
+                                min={0}
+                                step={1}
+                            />
+                        </div>
+                    ) : null}
+                </InspectorControls>
             ),
             ( <TextControl 
                 key="location-input"
@@ -117,11 +178,13 @@ export default class EditorBlock extends Component {
                     </p>
                 </div>
             ) : (
-                <div className={classNames}>
+                ( this.state.apiKey === '' && this.state.keySaved === false ) ?
+                keyInput
+                : (<div className={classNames}>
                     <div className="map">
                         {getMapHTML( attributes, this.state.apiKey )}
                     </div>
-                </div>
+                </div>)
             ),
 		];
